+++ internal/domain/account/fill.go
@@ -3,7 +3,7 @@ package account
 type Fill struct {
 	Symbol string
 	Side   Side
-	Qty    float64
+	Qty    int64
 	Price  float64
 	Fee    float64
 }
